+++ tiff-reader/src/tile.rs
@@ -508,7 +508,7 @@ fn read_tile_block(
     gdal_structural_metadata: Option<&GdalStructuralMetadata>,
 ) -> Result<Arc<Vec<u8>>> {
     let cache_key = BlockKey {
-        ifd_index: ifd.index,
+        ifd_offset: ifd.offset(),
         kind: BlockKind::Tile,
         block_index: spec.index,
     };